refactor: isolate strike merge verification (PAN-2907)
Issue: PAN-2907
