+++ scripts/file-size-baseline.txt
@@ -1,4 +1,4 @@
-1055 src/cli/commands/done.ts
+1028 src/cli/commands/done.ts
 1301 src/cli/commands/start.ts
 1412 src/cli/index.ts
 1012 src/dashboard/frontend/src/components/AwaitingMergePage.tsx
