+++ src/cli/commands/done.ts
@@ -42,6 +42,7 @@ import type { ScopeDriftRecord } from '../../lib/xbrief/continue-state.js';
 import type { XBriefDocument } from '../../lib/xbrief/types.js';
 import { hasOnlyPipelineStateChangesSinceCommit } from '../../lib/pipeline-state-paths.js';
 import { persistDoneReviewIntent } from './done-review-intent.js';
+import { verifyStrikeBranchMergedIntoMain } from './strike-merge-verification.js';
 
 const childProcessLayer = NodeChildProcessSpawner.layer.pipe(
   Layer.provide(Layer.mergeAll(NodeFileSystem.layer, NodePath.layer)),
@@ -67,10 +68,6 @@ interface SlotCompletionContext {
   slotIndex: number;
   slotItemId?: string;
   workspacePath: string | null;
-}
-
-function shellQuote(value: string): string {
-  return `'${value.replace(/'/g, `'"'"'`)}'`;
 }
 
 function parseSlotAgentId(input: string): { issueId: string; agentId: string; slotIndex: number } | null {
@@ -335,79 +332,6 @@ async function isMergeSetMergedIntoTargets(
   }
 
   return true;
-}
-
-export async function verifyStrikeBranchMergedIntoMain(issueId: string, projectPath: string): Promise<string> {
-  const branchName = `strike/${issueId.toLowerCase()}`;
-  const quotedBranch = shellQuote(branchName);
-
-  await execAsync('git fetch origin main', {
-    cwd: projectPath,
-    encoding: 'utf-8',
-    timeout: 60000,
-  });
-
-  let branchHead: string | undefined;
-  try {
-    const { stdout } = await execAsync(`git rev-parse --verify ${quotedBranch}`, {
-      cwd: projectPath,
-      encoding: 'utf-8',
-      timeout: 10000,
-    });
-    branchHead = stdout.trim();
-  } catch {
-    // A merged PR may delete the branch. The forge record below remains authoritative.
-  }
-
-  if (branchHead) {
-    try {
-      await execAsync(`git merge-base --is-ancestor ${quotedBranch} origin/main`, {
-        cwd: projectPath,
-        encoding: 'utf-8',
-        timeout: 10000,
-      });
-      return `${branchName} is contained in origin/main`;
-    } catch {
-      // Squash merges create a new commit, so the original branch is not an ancestor.
-    }
-  }
-
-  try {
-    const { stdout } = await execAsync(
-      `gh pr list --head ${quotedBranch} --state merged --json headRefOid,mergeCommit --limit 1`,
-      { cwd: projectPath, encoding: 'utf-8', timeout: 15000 },
-    );
-    const prs = JSON.parse(stdout) as Array<{ headRefOid?: string; mergeCommit?: { oid?: string } }>;
-    const pr = prs[0];
-    if (pr?.mergeCommit?.oid && (!branchHead || pr.headRefOid === branchHead)) {
-      await execAsync(`git merge-base --is-ancestor ${shellQuote(pr.mergeCommit.oid)} origin/main`, {
-        cwd: projectPath,
-        encoding: 'utf-8',
-        timeout: 10000,
-      });
-      return `${branchName} was merged by PR at ${pr.mergeCommit.oid}`;
-    }
-  } catch {
-    // No matching merged PR or forge access; fall through to local patch equivalence.
-  }
-
-  if (branchHead) {
-    const { stdout } = await execAsync(`git cherry origin/main ${quotedBranch}`, {
-      cwd: projectPath,
-      encoding: 'utf-8',
-      timeout: 10000,
-    });
-    const missingCommits = stdout
-      .split('\n')
-      .map((line) => line.trim())
-      .filter((line) => line.startsWith('+'));
-    if (missingCommits.length === 0) {
-      return `${branchName} has no commits missing from origin/main`;
-    }
-    throw new Error(`${branchName} has ${missingCommits.length} commit(s) missing from origin/main`);
-  }
-
-  throw new Error(`${branchName} is unavailable locally and no merged PR was found on origin/main`);
 }
 
 async function resolveDoneWorkspace(
